fix: Network select in explorer is showing kusama while on bsx
