+++ composables/usePrefix.ts
@@ -2,6 +2,8 @@ import { getKusamaAssetId } from '@/utils/api/bsx/query'
 
 export default function () {
   const { $store } = useNuxtApp()
+  const route = useRoute()
+  const router = useRouter()
 
   const urlPrefix = computed<string>(() => {
     return $store.getters.currentUrlPrefix
@@ -17,10 +19,20 @@ export default function () {
     return $store.getters['assets/getAssetById'](id)
   }
 
+  const checkPrefixBeforeMount = () => {
+    const prefix = route.params.prefix
+
+    if (urlPrefix.value !== prefix) {
+      $store.dispatch('setUrlPrefix', prefix)
+      router.go(0)
+    }
+  }
+
   return {
     urlPrefix,
     client,
     tokenId,
     assets,
+    checkPrefixBeforeMount,
   }
 }
