+++ pages/_prefix/detail/_id/index.vue
@@ -6,16 +6,8 @@
 import GalleryItem from '@/components/gallery/GalleryItem.vue'
 
 const { $store } = useNuxtApp()
-const { urlPrefix } = usePrefix()
-const route = useRoute()
-const router = useRouter()
 
-onBeforeMount(() => {
-  const prefix = route.params.prefix
+const { checkPrefixBeforeMount } = usePrefix()
 
-  if (urlPrefix.value !== prefix) {
-    $store.dispatch('setUrlPrefix', prefix)
-    router.go(0)
-  }
-})
+onBeforeMount(checkPrefixBeforeMount)
 </script>
