+++ pages/_prefix/gallery/_id/index.vue
@@ -5,17 +5,7 @@
 <script lang="ts" setup>
 import GalleryItem from '@/components/gallery/GalleryItem.vue'
 
-const { $store } = useNuxtApp()
-const { urlPrefix } = usePrefix()
-const route = useRoute()
-const router = useRouter()
+const { checkPrefixBeforeMount } = usePrefix()
 
-onBeforeMount(() => {
-  const prefix = route.params.prefix
-
-  if (urlPrefix.value !== prefix) {
-    $store.dispatch('setUrlPrefix', prefix)
-    router.go(0)
-  }
-})
+onBeforeMount(checkPrefixBeforeMount)
 </script>
